+++ src/components/Filters/FilterTypes/BrandFilters.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const BrandFilters = () => {
+  return <div>BrandFilter</div>;
+};
