+++ src/pages/Wishlist/Wishlist.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useCart } from "../../context";
+import { useCart } from "../../contexts";
 import { useScrollToTop, useDocumentTitle } from "../../hooks";
 import { ProductCard } from "../../components";
 import "./Wishlist.css";
